% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toddlr.R
\name{toddlr}
\alias{toddlr}
\title{Toddlr Shiny Gadget}
\usage{
toddlr(tw = toddlr::toc)
}
\arguments{
\item{tw}{tibble, output from rtweet containting thread: Default: toddlr::toc}
}
\description{
Shiny gadget for interactive dashboard
}
\details{
Thread self updates on load with lastest tweets
}
\examples{
\dontrun{
if(interactive()){
 toddlr()
 }
}
}
