% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toddlr_plots.R
\name{toddlr_plots}
\alias{toddlr_plots}
\title{Create toddlr output plot}
\usage{
toddlr_plots(plot_dat)
}
\arguments{
\item{plot_dat}{list of tibbles}
}
\value{
ggplot
}
\description{
Creates dashboard of plots from thread
}
\details{
Uses internal functions to create input list
}
