% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toddlr_slick.R
\name{toddlr_slick}
\alias{toddlr_slick}
\title{FUNCTION_TITLE}
\usage{
toddlr_slick(twe_dat)
}
\arguments{
\item{twe_dat}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rtweet]{tweet_embed}}
 \code{\link[slickR]{slickR}}
}
