% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threading.R
\name{tweet_threading_fw}
\alias{tweet_threading_fw}
\title{Forward update of twitter thread}
\usage{
tweet_threading_fw(tw, n = 1000, ...)
}
\arguments{
\item{tw}{tibble, rtweet output}

\item{n}{numeric, timeline of user to fetch to look for thread elements, Default: 1000}

\item{\dots}{Arguments to pass to \code{\link[rtweet]{get_timeline}}}
}
\value{
updated tibble, rtweet output
}
\description{
Function to update twitter thread from last status
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 tweet_threading_fw(toc)
 }
}
}
\seealso{
\code{\link[rtweet]{get_timeline}}
}
