% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toddlr.R
\name{toddlr}
\alias{toddlr}
\title{FUNCTION_TITLE}
\usage{
toddlr(tw = toc)
}
\arguments{
\item{tw}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
