% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toddlr_plots.R
\name{toddlr_plots}
\alias{toddlr_plots}
\title{FUNCTION_TITLE}
\usage{
toddlr_plots(plot_dat)
}
\arguments{
\item{plot_dat}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
