% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threading.R
\name{tweet_threading_fw}
\alias{tweet_threading_fw}
\title{FUNCTION_TITLE}
\usage{
tweet_threading_fw(tw, n = 1000, ...)
}
\arguments{
\item{tw}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION, Default: 1000}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rtweet]{get_timeline}}
}
